fastAPI/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
textblob
deep_translator